requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
python-multipart
